+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - modal: `scroll` option
   - script: `get(id)` function added in `script`execution context
   - script: `getProp(id, property)` function added in `script`execution context
+  - script: `id` variable added in `script`execution context (id of the widget that triggered the script)
 
 ## 0.46.1